first standard format
For the slug format I used `topic-prefix.key-terms`
For the description, this is a detailed description of the standard. Just text, like content.
For the assessment requirements, this is more of a guideline for how to decide what material gets tagged with the standard, and for us to know whether or not we have all the materials we need to assess the standard. Just text, like content.